Handle dragged files to topics directory
When files are dragged to topics directory, this violates the bulk
file syntax that every file must have a category that can be rendered
in the square brackets. The solution is to create a fake category and
indicate the user should edit the name.

A regex however in the message sorter assumed that all messages would
contain file paths with categories, and that assumption is removed.