Use unicode-bidi plaintext property to handle LTR-RTL mixes
This is the third commit in a sequence handling the introduction of
multiline text with padding between lines. This commit specifically
deals with the case of paragraphs of text that mixes lines that are LTR
and RTL. The desired behavior is for each line to be aligned correctly
without affecting the other lines.

The general approach to multiline text has changed three times.
- First, we added a text line matcher, but this broke multi-line tokens
  that begin midline, like bold and italic can.
- Second, we tried adding linebreak spans. This worked in general, but
  revealed the limitation regarding mixed RTL/LTR lines. To handle this
  case, we tried to identify which text spans were full-line and make
  them block and full width and dir-auto. This worked in general, but
  there were edge cases that it didn't address, like two sequential RTL
  links on one line, and identifying non-full-line spans of a text token
  that began or ended mid-line was very complicated because it required
  the parser which is largely stateless to know about tokens before and
  behind the current one. This would also not work for text blocks
  within other token types.
- The third approach was to use the `unicode-bidi: plaintext` option,
  which implements this desired behavior. However, Firefox has a bug
  that causes wrapping RTL text to sometimes wrap LTR, and so we needed
  a workaround involving requestAnimationFrame toggling the attribute
  to trigger a reevaluation of the text flow.

So, going forward, text tokens that contain newlines will be rendered as
multiple span elements per line, separated by .canopy-newline-span
elements. There is no terminal newline unless the text token included
a terminal newline literal as it might if followed by eg a multiline
token.