Allow mixed direction table lists
Previously table lists were either entirely left aligned or right
aligned, or centered. Now, we are supporting a use-case wherein a right
button is on the right and left on the left.

This commit also changes padding of table lists so that gaps are only
between buttons but not on ends.

Additionally, we are introducing a special case wherein a forward cycle
button which would normally get a forward arrow gets a backwards arrow
because of the semantic context ie going "backwards" even if to a path
that is not a subset of the current.

This commit also removes the margin between a block quote bar and the
given side of the page to align with text, so that text, block quotes,
and table lists will be horizontally aligned with one another.