Path reduction should center paragraph to make parent link visible.
Sometimes cycle links are clicked to go back to the mentioned topic,
but sometimes they are used for navigation as back buttons, in which
case the user really wants the parent link of the cycle root paragraph.

So, we will split the difference and put the cycle root paragraph lower
on the page so that the parent link is hopefully also visible.