Introduce fragment reference syntax
Previously, creating a link with an empty child paragraph allowed an
author to create a path reference pointing to the given parent link.

This commit introduces a new syntax to create a "fragment link," i.e.
a local reference to an empty subtopic not requiring the author create
the actual empty subtopic.

The syntax [#[ABC]] will create a local link with the given subtopic
name, and will create an empty subtopic paragraph for that subtopic.