Consolidate scroll functions
Previously, path changes were modeled as a change of path followed by
a scroll to the new selected paragraph, unless there were conditions
necessary for an animation, i.e. a pre-path-change scroll e.g. in order
to put a fulcrum element in view before proceeding to open a different
path below, etc.

This logic is now replaced with a simpler formulation wherein every
displayPath may do a beforeChange scroll and/or an after change scroll.
A beforeChange scroll either puts a much higher element or fulcrum
element into view, and then an afterChange scroll puts the given
paragraph or parent link in view.