Text link button redesign
Previously links in text were underlined in accordance with general
web standards. However, this made situations where e.g. a "more" button
was being created a bit awkward, because fundamentally it was just
underlined text, not a "button" in the standard UX metaphor.

So, now text links are being styled similar to how menu elements and
table cells already were, with a border indicating clickability and
different hover, selection, and open states indicated by colors.

Creating a background color to text links breaks an earlier solution to
preventing punctuation from wrapping after links, wherein we had been
using a padding plus negative margin, but which would now be background-
colored. So, now links have a content container which holds the color
styling, and the A tag itself is a wrapper that can have the padding and
negative margin without affecting color.