Indicate scroll preference with optional link argument
UpdateView has formal parameters of a path and link argument. However,
never are both required. By passing a link, we can determine from its
type what path should be visible when it is selected, unless it is an
inlined cycle link, in which case by passing the inline cycle path we
clarify the link's open state.

Previously, we had populated the missing argument from the present one,
either in the caller or in the function body eg first line of
displayPath. However, semantically this obfuscated whether the caller
itself had specified eg a link argument, or whether the path was
the significant component.

This commit also removes the concept of scrollToParagraph as an
argument, replacing it with the aforementioned presence or non-presence
of linkToSelect as the representation of the caller's intention to
scroll to link or paragraph.

A scrollToParagraph argument is preserved in order to allow link#execute
to remain initiator-neutral (i.e. keys vs. mouse), however, the flag
is just a wrapper for link#select to defer to path#display.