Support link icons for RTL text
Previously we had assumed that external and cycle link arrows would
always go on the right side, but this wasn't true and caused
irregular link border behavior.

Now, links get a dir=auto which causes icons to be rendered on the
correct side. We also put padding/negative margin on the left side
to prevent punctuation from being separated in RTL text also.