Adjust link borders for diacritical marks
Previously borders were given by the bounding box of the text, however,
at times the text would overflow this and it wasn't detectable via the
standard DOM API.

Now, we take each link text and draw the text on an invisible canvas
element so that we can detect the exact vertical overflow by comparing
to a full-height character like ( and then we add that amount as padding
to the top or bottom of the element.

This commit introduces a helper to wait for token elements to be
inserted into the DOM before doing certain calculations.