Initial page load should scroll to paragraph
Previously initial page load focused on parent link of selected
paragraph to make arrow key navigation easier and to show context of
current content. However, it was confusing that the most recent scroll
state e.g. before refresh would be different from after refresh, and
in general the user probably wants the target paragraph more than the
parent link. Path references where we may choose to focus on parent link
is a special case where the link author indicated interest in the path
to a given paragraph.

This change requires many tests to be changed to have a scroll before
a given keypress because now the selected parent link isn't always
visible on page load.