All upward path movements should scroll first
Previously any 1-step path change would change and then scroll, and
we would only scroll first in the distance up was very long, or it was
a two-step movement. However, it looks better for the thing in focus
not to disappear, even when the new target is already visible. This
simplifies the logic and obviates the need to detect "large" scrolls.