Executing topic link in topic paragraph pops paragraph stack
Previously a link in a subtopic to the enclosing topic produced a cycle,
the reduction of which resulted in an effective back scroll to the topic
subtopic. However, there was never a good way for the author to create
a similar back button effect from the topic subtopic itself, i.e. as if
to make a global link to the subtopic prior on the stack, the reduction
of which cycle would effectively pop the topic subtopic from the stack.

At the same time, there was always a question what a global link to
a topic should do if present in the topic subtopic of that topic. The
user is already looking at the target, and without an exception, the
default behavior would be to reduce the cycle and create a no-op.
Previously we had detected this case and forced inlining so as to make
it do something, even though this wasn't likely to be a realistic
use case.

Now however, we can adapt this command of self-topic-reference to a much
more plausible authorial intent, i.e. appropriate it for use as a
trans-topic back button, equivalent eg to clicking an open and
selected parent link causing it to toggle closed. This change allows
authors to make a simple back button from a topic subtopic for e.g.
simple background informational topics.

At present there exists no way for an author to create a stack-popping
link from within a non-topic-subtopic, but this effect can be achieved
by making a subtopic back button combined with a topic-subtopic back
button.