Create unique behavior for self, rehash, and coterminal references
Since the introduction of path reduction behavior, execution of
references that introduce cycles causes the cycle to be removed.

Previously we observed that a link to a topic in that topic's subtopic
paragraph, e.g. in A#A a link [[A]], is a no-op because the cycle A/A is
immediately reduced. So, we repurposed such a link to mean a request to
pop A from the current stack, i.e. a back button, to make eg B/A go back
to B, the same way that links to a subtopic's enclosing topic have
already been used as automatic back buttons to the topic root, taking
advantage of the default cycle reduction behavior.

Now, we note that a topic self-reference e.g. [[T]] in topic T's root
paragraph is merely an example of a larger category of link which are
equally non-operative, called coterminal references.

A coterminal reference is any reference that in a given explanation
reiterates some portion of the end of the current path.

For example, if the path is A/B/C/D/E, then if paragraph E contained
one of these references: [[E]], [[D/E]], [[C/D/E]], [[B/C/D/E]], or
[[A/B/C/D/E]], these would all be no-op coterminal references.

Additionally, given the path A/B/C/D/E, we can imagine other coterminal
references, for example: [[X/Y/Z/E]] or [[X/Y/Z/D/E]] i.e. appearing in
paragraph E. These references begin with a different context and then
coincide with a terminal portion of the current path. These paths are
also non-operative, for the same reason as the first set, and all are
therefore open for repurpose.

We can break this larger category of "coterminal references" into three
distinct categories.

First, the simplest, e.g. paragraph A/B/C containing a link [[C]],
can be called self-references. Another example would be A/B#C containing
a reference to [[B#C]]. These are unique because they contain only one
path segment, and so we don't see a clear intention of the author to
identify a path, rather all we see is the extra emphasis of a given
paragraph, and so the logical behavior to associate with it would be to
remove the given paragraph.

How exactly should self-references work? Should they "pop" i.e. make
A/B/C go to A/B, or should they merely scroll up to the current parent
link i.e. "shift"? We will follow the following approach:

+-------+-----------------------------+------------------------------+
|       |  From within a ST paragraph |   From within a T paragraph  |
+-------+-----------------------------+------------------------------+
|  Pop  |    Parent reference [[^]]   |  Self-reference [[^]] [[.]]  |
+-------+-----------------------------+------------------------------+
| Shift |     Self-reference [[.]]    |       Rehash Reference       |
+-------+-----------------------------+------------------------------+

Topics have no way of referencing their parent because it isn't known
at compile time, so a self-reference is the only way an author would
have to express a topic paragraph should fully pop. Conversely,
non-topic subtopics already have the [[^]] syntax which creates a
reference to the parent subtopic, which via normal cycle reduction
produces pop behavior. So, self-reference is available in non-topic
subtopic paragraphs to indicate the shift behavior. For shift behavior
in topic paragraphs, authors can use the rehash syntax discussed
below.

The second category of coterminal references we mentioned can be called
rehash references, e.g. A/B/C containing a reference [[B/C]].

For rehash references, by rehashing a portion of the prior path, the
author seems to be emphasizing the path to the current paragraph, and
so it seems like a reasonable response would be to scroll upwards to
the first parent link in the rehash reference path, e.g. if in topic
New York, a link to [[United States/New York]] would scroll up to
United States's link to New York.

For merely coterminal references, e.g. A/B/C containing a reference[
[E/F/B/C]], by writing a non-operative reference that adds a new
contextual prefix, the author seems to be emphasizing the existance of
an alternative path to the current location. Given this, it seems
appropriate to inline rather than reduce the cycle, and to scroll to
the parent link in the inline path at which point the new path diverges
from the old. For example:

At path A/B/C, if we had reference [[E/F/B/C]], we would inline the
reference to display A/B/C/D/E/F/B/C, and then would focus on Y's link
to B, because this is the point at which the new introduction
transitions to the current discussion.

Because this behavior is dependent on current path, an author can use
coterminal references to create conditional behavior - if the reference
rehashes the path by which the reader did in fact arrive at the current
content, then it should regress, and if the reader arrived there some
other way, the referenced path should be mentioned in addition i.e.
inlined.