Simplify beforeChange scroll logic
When a user selects a path to display, there are several directions in
which that new path may be relative to the current one. It might be a
full-redirect with no continuity to the current page. It might be a
simple down movement, eg A/B/C to A/B/C/D. It might be a simple up
movement, eg A/B/C to A/B. Until recently, large ups were a separate
case eg A/B/C/D to A/B, as this gave more reason to scroll before path-
changing, although now all upward movements scroll before path-change.
You could have an upward movement that is large visually but is only
one subtopic pop, or an upward movement that is going up multiple
subtopics. Lastly, you have a two-step change, e.g. from A/B/C to
A/B/D.

The question has been, in which of these circumstances should there be a
pre-change scroll. Until this commit, the list was as accumulation of
different empirical cases, whereas this commit introduces comprehensive
criteria.

  1. All downward movement changes before scroll so that the scroll is
     to a populated space.

  2. All upward movement scrolls before changing so that focused content
     isn't erased.

     However, unlike downward movement where a path change revealing
     content always necessitates a downward motion to focus, for upward
     motion, the user may coincidentally already be focused on the
     destination of the theoretical pre-change scroll.

     As such, after determining the upward direction of the path change,
     we must empirically detect the user's scroll position to determine
     if there should be a pre-change scroll.

So, now we move all the gating logic into beforeChangeScroll, and can
get rid of the complicated twoStepChange function and replace it with
a check that the paths are overlapping and that their fulcrum element is
not focused. Rather than trying to calculate by looking at the path
how large the first movement would have to be, we can use the focus
state of the fulcrum element to obviate such approximations.