Introduce multi-line button link styling
With the addition of button link styling, the situation emerged that a
link might span multiple lines, and each line would receive its own
border. This was fine if it was inline text, but sometimes the link was
entirely alone on a line. In such a case, it would be more appropriate
to draw a border around the entire multi-line element, rather than
around each line specifically.

This feature introduces the possibility of producing trees of prose
in which the link is not a certain word or phrase being referenced,
indicating a definition or specification of just that word, but rather
allows the linkification of a full section of the text itself. This
perhaps will allow the UX shift of implying a specification of the
entire text portion, rather than just of some word within it, which may
enable trees of specificatory content.

This commit also removes the #_canopy container from the Sass rules,
because it sometimes breaks sibling selectors, eg:

\#_canopy {
  a {
    .canopy-linebreak-span + &:last-child {
       display: inline-block:
    }
  }
}

This compiles to:

.canopy-linebreak-span + \#canopy a:last-child { ...

I.e, it expects the container to be the sibling of
.canopy-linebreak-span which of course never happens. Removing the
outer container means that most of the time styles are being applied
to elements that can be moved around as such in a nested rule.