Use spans for link icons rather than ::after blocks
This change transitions from using ::after blocks for various post-link
icons to using regular spans.

This was motivated by a desire to improve the accuracy of menu sizing
which has great difficulty detecting and estimating the size of ::after
blocks. However, the change also improves the legibility of tests,
asserting on visible text, and simplifies the DOM structure in general.