Bug: Only give margin to inner-most left or right aligned menu cell
When multiple menu cells are right or left aligned and there is at least
one of each alignment, only the inner-most cell of each group should
have a margin auto, otherwise the cells of a group will be separated.

We need to maintain the original CSS class because we check this e.g.
when flipping the arrow icon for a lateral cycle menu link that is left
aligned.