Bug: Fragment links should override click handler scrollToParagraph behavior
When clicking a link, we generally scroll to paragraph. The scroll
helpers knew that for fragment links we scroll to the link e.g. when
displaying the path of the fragment child, but when a scrollToParagraph
true argument was passed, this was overriding that desired behavior,
so now scrollToParagraph is overridden by Link#isFragment.