Bug: Consider \n character in calculating blockquote padding
When rendering blockquotes, we try to detect wrapping of text because
if some linebreaks are wraps and some are semantic, we want to add
padding to the real linebreaks to differentiate.

However, previously we were only looking for canopy-linebreak spans
to determine which linebreaks were intentional, however, if e.g. the
user creates an HTML block that spans multiple lines but without
wrapping, we weren't detecting the \n characters as signs of intentional
linebreak and were considering them wraps to trigger padding.

Now we consider BR, canopy-linebreak-span, and regular \n in a textnode
to be intentional linebreaks, as opposed to wraps.