Clarify logic for pauses in click-initiated two-step changes
+-------------+-----------------+--------+--------------+--------------+
| Scroll up   | First Pause     | Change | Second Pause | Scroll Down  |
+-------------+-----------------+--------+--------------+--------------+
| If logical  | Only if first   | Always | If logical   | If necessary |
| two-step    | scroll actually |        | two-step     |              |
| and fulcrum | occurred        |        | and fulcrum  |              |
| not focused |                 |        | was not      |              |
|             |                 |        | clicked      |              |
+-------------+-----------------+--------+--------------+--------------+

Click handlers have passed various options into Link#execute such as
scrollToParagraph and noBeforeChangeScroll, the latter because the point
of a pre-change scroll is to focus the user on the fulcrum, but if they
clicked a global link ie fulcrum, then they are aware of the fulcrum
already etc.

However, when the user clicks a cycle link that reduces, the fact that
it was a click that initiated the change does not imply the user has
an awareness of the fulcrum, which in that case is no longer the clicked
link.

Indeed, when a cycle is being reduced, almost none of the options passed
having to do with it being click-initiated are now relevant as the
current change is being caused by the cycle reduction, not the click
directly.

Additionally, this commit cleans up some logic around before- and after-
change pauses, and when they should be omitted.

The before-change pause is there to to stop movement before changing
link selection so as to isolate visual changes, but if the fulcrum was
already focused, then there was no actual beforeChange scroll, and so
this pause is unnecessary.

The after-change pause exists to give a moment of emphasis to the
selection of the fulcrum. This pause therefore is not dependent on there
having been a real first scroll, but should occur so long as the
total path change contains two logical steps, ie they have a merely
partial initial overlap, necessitating a logical move up and a move
down, whether or not those moves were realized as actual scrolls.
Additionally, if the change was initiated by a click of the fulcrum
element, this pause is unnecessary because the user has already
expressed familiarity with the new fulcrum selection.

Additionally this commit fixes a bug where previously the delay promises
were created before use which meant that they had resolved before being
used and didn't cause a delay.