Lessen padding under DOM content
We add space after the content so that if the user navigates upwards,
the removal of content from the DOM doesn't cause the bottom of the DOM
to become visible which cancels the scroll. However, we don't want
so much empty space down there, and now anyway scroll precedes content
change for all upward motion so this error case is unlikely.