In bulkfile generation, prefer shorter of topics with category substring
Sorting topics of a category in the bulk file in a way that balances
alphabetization but also logical ordering of topics which may have an
inherent sequence is challenging. Primarily, users who want logical
ordering can either create a numeric topic naming scheme that preserves
alphabetization, or they can create dummy categories like /1/X /2/Y,
etc.

A third option was previously introduced wherein a topic that contained
its terminal topic name as a substring got hoisted to the top.
However, sometimes there were multiple topics that all contained the
category as a substring. Now, we apply the rule that the shortest of
these is probably functioning as a table of contents for the others.