Rename .canopy-arrow-keys-container to reflect general use
Renderers may apply the .canopy-bounding-box-container to an element
that may contain a .canopy-selectable-link if and when were a link to
exist as a descendant, the ancestor element's borders should represent
the link's for all geometric purposes. The class was originally named
for the use-case of arrow key calculations, but it also affects things
like positioning links at a certain percentage of the viewport, etc,
so the new name reflects its more general purpose.