Improve handling of fragment references when using --cache option
Previously using a fragment reference had caused the final JSON to
include fragments, but not the temporary internal values of the
parserContext object. However, errors could be caused by validation
not knowing about these new subtopics. So, now we do add them to the
internal stores and also only validate on the non-cached run.