Distinguish "down cycles" that should not get arrow icons
Previously cycles were divided into two categories, back cycles e.g.
a link to [B] in A/B/C which reduced would redirect to A/B, or lateral
cycles e.g. a link to [B/D] in A/B/C which would go up to B and add D.

However, there is a third logical category which had been classified as
lateral references but which should have individual behavior. These are
now called "down cycles."

An example of a down cycle is as follows. A paragraph A may contain
a variety of links e.g. to B and C. Someone may which to create an
additional link that suggests where to start. They may create a link
[[A/C]], which when clicked, reduces to A/C, and the parent link becomes
the simplest reference in A to C, i.e. the original link to C, not the
helper link that was clicked. Thus, the result of selecting the helper
link is to select the other link C, resulting in an entirely downward
motion so long as C is not much higher than the helper.

Thus, it doesn't seem appropriate for such a link to have a forward
arrow icon as do other lateral links.