Link border adjustments should add 1px to fitting bottoms
It doesn't look good for a border to be exactly at the end of the text,
so exact fits get a 1px padding. Zero diff above doesn't require padding
because even if the link text is the same height as our reference
string "()", nevertheless even text with 0 padding gets padded by the
text's line height, which sets the minimum border, so we just don't add
on top.

This commit refactors external links and fragment links to use the
linkBase helper.