Update history functions to allow link vs. paragraph selection state
Recent commits have introduced the convention that calls to updateView
that contain a link argument signify a request to scroll to that link,
whereas if the link argument is inferred from the path argument, we
want to scroll to the path's target paragraph, unless overridden by the
scrollToParagraph option.

However, history states and the corresponding popstate listener had not
been updated to use this new format, making it impossible e.g. to store
a history state that reflected a link selection versus a paragraph
selection.

Additionally, the use of options.pushHistoryState had come to comprise
two separate uses, first, clarifying that a setPath should push rather
than replace the existing state, and second, that certain interactions
such as clicking a cycle link should push _an additional_ history
stack frame to represent the intermediary stage of cycle link selection
between the prior path and the reduced cycle path. Now, we have removed
the pushHistoryState option, and instead, for setPath we rely on set
logic for whether a new path is different from the previous without
additional options, and situations where an extra stack frame should be
pushed are handled by the existing Link.pushHistoryState.