Assign cycle link arrows based on relative direction of paths
Links whose execution would introduce a cycle have been given arrow
icons to inform the user of the different execution behavior.

Initially, there were only back cycle links, i.e. those which took
a user strictly upwards, and "lateral," which took a user either
forward or backwards in the path stack.

However, this meant that cycle links which took a person backwards
had a forward arrow, counter to user expectations.

(Additionally, between then and now the idea of a "down" cycle was
introduced i.e. a strictly down motion e.g. [[B/C]] in B that doesn't
get an arrow icon.)

So, this commit introduces a check whether the link's reduced inline
path ie target would be a forwards, backwards, upwards, or downwards
movement relative to the current, and assigns icons accordingly.

In implementing this change, the primary obstacle is the fact that many
of these comparisons occur on initial page render, at which time the
link's target path may not yet be rendered. In theory this shouldn't
matter because given any two initially-overlapping paths, one of which
is rendered, the "paragraph of departure" which includes both parent
links will always be visible, and this is all we need to determine
comparative position in DOM. However, many of the helpers that deal with
parent links assume that we can access path -> paragraph -> parent link,
however, it is possible for the overlap paragraph to be rendered and yet
the link's target paragraph to not be, in which case this lookup will
fail. So, many helpers had to be rewritten to detect the path -> parent
link lookup and if the last path segment is a topic segment, to go
looking for the parent link in the path.withoutLastSegment that matches
the last segment, without looking for the paragraph associated with the
last segment.