Link selection state should only be pushed if non-redundant
Previously we introduced the behavior of pushing a selection state
stack frame for link executions where the result left the user away from
the given link, and a browser back should therefore go back to the
executed link before the prior state.

However, this could produce the following scenario:

- In A, click [[B]] (history stack: Paragraph A, [[B]], Paragraph B)

- In B, click [[C]] (history: Paragraph A, [[B]], Paragraph B, [[C]])

Now, a browser back will go from [[C]] to Paragraph B, but [[C]] is in
B, and so the visual distance is not large, and the two stack frames
will seem redundant.

In the event that a push of a new link selection would be very similar
in y-coordinate to an existing non-link paragraph selection, now we
do nothing and allow the earlier paragraph selection to represent both.
(As opposed to updating the stack frame to reflect the new link, because
it is more likely the user wants to return to a paragraph top if both
are proximate.)

One final case is if a link selection state is being pushed and it is
likely that a paragraph state will be pushed soon after for the
paragraph immediately succeeding the link, if that link is close to the
bottom of its paragraph, we will omit the push of the link state and
allow the subsequently pushed paragraph state represent that
chronological page position.

So, two times we intervene and do not push a link state:

1. If the link is close to the previous paragraph state

2. If the link state is likely to end up being close to the subsequent
   paragraph state.