Chunk plaintext to improve performance of parser
Previously, we would take a given substring and try it against all of
the various matchers for special tokens, and if we didn't find any
matches, then we would remove the first character as plaintext and try
again on the remaining string.

However, although we cannot tell which tokens exist without running
all the matchers, we can sometimes know that no token will be triggered
on some substring due the absence of any special characters or character
sequences.

If so, rather than running all the matchers on every substring, we can
add plaintext to the buffer so long as we continue to see only plaintext
characters.

On a large sample project, this change seems to reduce build times by
50%.