Unicode normalize paragraph keys and link targets.
A link target and topic name can be identical to the end user, but
be composed of diacritical characters that on disk are in different
orders.

Now, we standardize the orderings so that a character which is identical
to the user is treated as identical for link resolution purposes.

Standardization is done at the mixed case form so that file names like
JSON are also standardized.

Changes are not necessary to the bulk mode parser because the topic file
names that come from differing display keys will be standardized via the
mixed case form.