Restore category notes behavior in Canopy bulk mode
Canopy bulk mode allows a user to bulk create and edit multiple expl
files by editing one bulk file.

Paragraphs i.e. \n\n separated text in the bulk file is interpreted
as a note, i.e. a provisional pre-topic placeholder that is not parsed
or offered in the web client.

When the user makes notes in the bulk file, there is a question which
file to put them in. Initially the idea had been that for any category
that contained notes in the bulk file, we would create an expl file
by the name of that category to put them in, and that if one already
existed i.e. a topic by the name of that category i.e. a category topic,
then we would append any notes from the bulk file to the end of that
existing file.

When we were considering having ordered topics and prefixing topic file
names with numbers, we moved away from this approach and began making
each note into its own note file, etc.

Now that we are preparing to introduce the Canopy Review feature, it
seems a revision to the original behavior would be most appropriate. In
Canopy Review, users will create collections of notes which they review
and revise in order to condense, subdivide, and make laterally
consistent between notes. For this process, it is most helpful for notes
of a category to be consolidated in one place, because anyway the goal
is to condense and integrate them into an explicit prose structure.

So, we are reverting the more recent changes and returning to the
category notes file behavior.