Fix playground error
Canopy.js assumes a #_canopy element exists in the DOM before
initialization. `getters.js` is the first point of failure if it does
not exist.

We use REQUEST_CACHE to store HTTP promises across different parts of
the code.

A recent feature added a reference to getters in REQUEST_CACHE, so as to
initialize the cache with certain values pre-loaded. However, in
canopy consumers like the playground code, the REQUEST_CACHE can be
loaded before page initialization to stub data for HTTP requests, and
so the new call to getters.js broke that code.

Now, the playground code waits to load REQUEST_CACHE until the DOM is
initialized. Another solution which may be implemented in future is to
have Canopy fail more gracefully when run on an unready DOM, e.g. trying
again on some setTimeout interval, etc.

This commit also adds testing over the playground, and caches the
playwright browser download from the earlier test suite.

This commit also parallelizes the playwright tests to facilitate quicker
builds.