Escape colons in bulk mode category notes
If in bulk mode we have a category note with a colon, e.g:

```
[Category]

This is: a note.
```

Which is distinct from the user deciding to create a topic:

```
[Category]

* This is: a topic.
```

Then we want to save the former in such a way that we retain the fact
that it was not made a topic by its author. However, saving it with a
colon in the first paragraph before other punctuation is the way topics
are stored from the bulk file, i.e. the result of the second syntax
above also.

So, now we escape the colon of the former, to distinguish it from the
latter, as apparently we had already been doing with question marks.