Bug: Root topic load should focus top of paragraph
We were looking for Path#firstTopicPath on the class when it is an
instance method.

Now we check if the path to display is the root paragraph and if so, we
don't try to put it at 30% of viewport because this will hide the header
text, which is probably not what the user wants specifically when it is
the root paragraph that is being displayed.